pyproject.toml
[project]
name = "zeroth-core"
version = "0.1.0"
description = "Governed medium-code platform for production-grade multi-agent systems"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "governai @ git+https://github.com/rrrozhd/governai.git@7452de4db144e8c92b7b5ae0a646c40314a6aa99",
    "fastapi>=0.115",
    "httpx>=0.27",
    "langchain-litellm>=0.3.4",
    "litellm>=1.83,<2.0",
    "PyJWT[crypto]>=2.10",
    "pydantic>=2.10",
    "pydantic-settings>=2.13",
    "psycopg[binary]>=3.3",
    "psycopg-pool>=3.2",
    "aiosqlite>=0.22",
    "alembic>=1.18",
    "PyYAML>=6.0",
    "python-dotenv>=1.0",
    "sqlalchemy>=2.0",
    "econ-instrumentation-sdk @ file:///Users/dondoe/coding/regulus/sdk/python",
    "cachetools>=5.5",
    "redis>=5.0.0",
    "tenacity>=8.2",
    "uvicorn>=0.30",
    "pgvector>=0.4.2",
    "litellm>=1.83.0",
    "psycopg[binary]>=3.1",
    "chromadb-client>=1.5.6",
    "elasticsearch[async]>=8.0,<9",
    "arq>=0.27",
    "mcp>=1.7,<2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/zeroth/core"]

[tool.hatch.metadata]
allow-direct-references = true

[dependency-groups]
dev = [
    "libcst>=1.4",
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "ruff>=0.11",
    "testcontainers[postgres]>=4.14",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-m 'not live'"
markers = [
    "live: marks tests requiring live API keys (deselect with '-m not live')",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
